Create on PersistentEntitytestDriver per @Test. (#25)
The test execution in sbt is creating one instance for each @Test in
the test class which causes the PersistentEntitytestDriver `driver`
instance to be different on each @Test. The test, as previously
written gave the impression that PersistentEntitytestDriver provided by
Lagom is stateless and that a single instance can be reused between
tests which is not true.